refactor: HttpInterface 사용하여 API 호출 직관적으로 구조화
