+++ src/main/java/roomescape/dto/payment/PaymentRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto.response;
+package roomescape.dto.payment;
 
 public record PaymentRequest(String orderId, int amount, String paymentKey) {
 }
