+++ src/main/java/roomescape/dto/payment/PaymentResponse.java
@@ -1,0 +1,4 @@
+package roomescape.dto.payment;
+
+public record PaymentResponse(String paymentKey, int totalAmount) {
+}
