+++ src/main/java/roomescape/service/PaymentService.java
@@ -1,7 +1,14 @@
 package roomescape.service;
 
-import roomescape.dto.request.reservation.ReservationRequest;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.service.annotation.HttpExchange;
+import org.springframework.web.service.annotation.PostExchange;
+import roomescape.dto.payment.PaymentRequest;
+import roomescape.dto.payment.PaymentResponse;
 
+@HttpExchange
 public interface PaymentService {
-    ReservationRequest pay(ReservationRequest reservationRequest);
+
+    @PostExchange("/confirm")
+    PaymentResponse pay(@RequestBody PaymentRequest paymentRequest);
 }
